+++ src/daf/logging/sql/tables.py
@@ -14,6 +14,8 @@ from sqlalchemy.orm import (
     DeclarativeBase
 )
 
+from ...misc.instance_track import track_id
+
 
 class ORMBase(DeclarativeBase):
     "Base for all of ORM classes"