+++ docs/source/changelog.rst
@@ -37,6 +37,19 @@ Glossary
 Releases
 ---------------------
 
+v3.1.0
+===================
+- JSON Analytics
+- GUI:
+  
+  - ViewOnly structured data will display only the data that is provided, meaning
+    the GUI will not be constructed based on type annotations of an objects, but rather
+    based on the data itself.
+
+- LoggerJSON ``index`` field of logs is now an unique snowflake like id.
+- |BREAK_CH| :class:`daf.logging.LoggerJSON`'s invite logs will now contain a "member" dictionary
+  for each invite log.
+
 v3.0.4
 ====================
 - Fixed AutoGUILD not working if the ``messages`` parameter is None.